+++ sysdeps/unix/sysv/solaris2/kopensolaris-gnu/sendfile.c
@@ -31,7 +31,7 @@ sendfile (int out_fd, int in_fd, off_t *offset, size_t count)
   sfv.sfv_len = count;
   int xferred = 0;
 
-  int res = sendfilev (out_fd, &sfv, 1, &xferred);
+  ssize_t res = sendfilev (out_fd, &sfv, 1, &xferred);
   *offset += xferred;
   return res;
 }
